feat: added benchmark script
